+++ openghg/standardise/surface/_icos.py
@@ -217,6 +217,7 @@ def _read_data_large_header(
         "network": network,
         "instrument": instrument,
     }
+    attributes = {"inlet_height_magl": metadata["inlet_height_magl"], "data_owner": "See data_owner_email"}
 
     if measurement_type is not None:
         metadata["measurement_type"] = measurement_type
@@ -246,12 +247,7 @@ def _read_data_large_header(
         if interval_str == "hourly":
             metadata["sampling_period"] = "3600.0"
 
-    species_data = {
-        species_fname.lower(): {
-            "metadata": metadata,
-            "data": data,
-        }
-    }
+    species_data = {species_fname.lower(): {"metadata": metadata, "data": data, "attributes": attributes}}
 
     return species_data
 
@@ -377,15 +373,11 @@ def _read_data_small_header(
         "data_type": "surface",
         "source_format": "icos",
     }
+    attributes = {"inlet_height_magl": metadata["inlet"], "data_owner": "NOT_SET"}
 
     if measurement_type is not None:
         metadata["measurement_type"] = measurement_type
 
-    species_data = {
-        species_fname: {
-            "metadata": metadata,
-            "data": data,
-        }
-    }
+    species_data = {species_fname: {"metadata": metadata, "data": data, "attributes": attributes}}
 
     return species_data