+++ tests/standardise/surface/test_icos.py
@@ -30,6 +30,7 @@ def test_read_icos_large_header():
         "station_latitude": 51.99747,
         "station_long_name": "Ridge Hill, UK",
         "station_height_masl": 207.0,
+        'data_owner': 'See data_owner_email',
     }
 
     metadata = data["ch4"]["metadata"]
@@ -110,6 +111,9 @@ def test_read_icos_small_header_file():
         "station_latitude": 56.55511,
         "station_long_name": "Angus Tower, UK",
         "station_height_masl": 300.0,
+        'data_owner': 'NOT_SET',
+        'inlet_height_magl': '222m',
+
     }
 
     assert attrs == expected_attrs
@@ -136,6 +140,8 @@ def test_read_icos_small_header_file():
         "station_height_masl": 300.0,
         "data_type": "surface",
         "source_format": "icos",
+        'data_owner': 'NOT_SET',
+        'inlet_height_magl': '222m',
     }
 
     assert co2_metadata == expected_metadata